Integrated RUIS and changed camera
The RUIS integration requires I use the RUIS object, which has a
DisplayManager. To use the DisplayManager for the Oculus, I have to use
RUISCamera's instead of unity camera for my display. There are also some
differences in how I have to switch cameras because the Display under the
DisplayManager requires a connected camera.